add grid view
