+++ backend/graminsta/post/urls.py
@@ -7,7 +7,6 @@ from django.urls import path
 from .views import (PostRecordView,
                     FollowView,
                     PersonalGalleryView,
-                    PersonalCountView,
                     TimelineView)
 
 
@@ -17,6 +16,5 @@ urlpatterns = [
     path('follow/', FollowView.as_view(), name='follow'),
     path('follow/<user_id>/', FollowView.as_view(), name='following_people'),
     path('timeline/', TimelineView.as_view(), name='timeline'),
-    path('personal/', PersonalGalleryView.as_view(), name='personal_gallery'),
-    path('personal_count/', PersonalCountView.as_view(), name='personal_count')
+    path('personal/', PersonalGalleryView.as_view(), name='personal_gallery')
 ]
