+++ flutter/Graminsta/lib/constants.dart
@@ -1,1 +1,4 @@
 const userTokenKey = 'user_token';
+const postCountKey = 'post_count';
+const followingCountKey = 'following_count';
+const fanCountKey = 'fan_count';
