+++ flutter/Graminsta/lib/spec/spacing.dart
@@ -1,1 +1,2 @@
 const double authTopSpacing = 200;
+const double personalGalleryImgPadding = 1.0;
